requirements.txt
pandas
numpy
scikit-learn
lightgbm
scipy
joblib
nltk
flask
spotipy
python-dotenv